requirements.txt
# FastAPI 核心
fastapi>=0.104
uvicorn[standard]>=0.24

# 数据库
SQLAlchemy>=2.0
alembic>=1.8
pymysql>=1.1.0
cryptography>=41.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 缓存
redis>=5.0

# 数据处理
requests>=2.28
pandas>=1.5
beautifulsoup4>=4.11
lxml>=4.9
python-dotenv>=0.21